+++ packages/client/signers/src/communications/index.ts
@@ -5,3 +5,5 @@ export type {
     SignerInputEvent,
     SignerOutputEvent,
 } from "./events";
+
+export type { KeyType, Encoding, KEY_ENCODINGS, KEY_TYPES } from "./schemas";